requirements.txt
tensorflow-cpu==2.16.1
numpy>=1.21.0
Pillow>=9.0.0
Flask>=3.0.0
streamlit>=1.24.0
